Backend/requirement.txt
Flask
Flask-Session
Flask-Bcrypt
psycopg2
cohere
python-dotenv
Flask-Cors
